# quiz/swagger.json
{
   "swagger":"2.0",
   "info":{
      "description":"\"Quiz Application provide question and answer and display the results\"",
      "title":"Quiz Application"
   },
   "host":"localhost:8080",
   "basePath":"/",
   "tags":[
      {
         "name":"quiz-controller",
         "description":"Quiz Controller"
      }
   ],
   "paths":{
      "/quiz/v1/user/getFinalResult":{
         "post":{
            "tags":[
               "quiz-controller"
            ],
            "summary":"getFinalResult",
            "operationId":"getFinalResultUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"request",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            }
         }
      },
      "/quiz/v1/user/getFirstQuestion":{
         "post":{
            "tags":[
               "quiz-controller"
            ],
            "summary":"getFirstQuestion",
            "operationId":"getFirstQuestionUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"request",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            }
         }
      },
      "/quiz/v1/user/getNextQuestion":{
         "post":{
            "tags":[
               "quiz-controller"
            ],
            "summary":"getNextQuestion",
            "operationId":"getNextQuestionUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"request",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            }
         }
      },
      "/quiz/v1/user/getPreviousQuestion":{
         "post":{
            "tags":[
               "quiz-controller"
            ],
            "summary":"getPreviousQuestion",
            "operationId":"getPreviousQuestionUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"request",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/QuestionAnswerResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            }
         }
      },
      "/quiz/v1/user/login":{
         "post":{
            "tags":[
               "quiz-controller"
            ],
            "summary":"login",
            "operationId":"loginUsingPOST",
            "consumes":[
               "application/json"
            ],
            "produces":[
               "*/*"
            ],
            "parameters":[
               {
                  "in":"body",
                  "name":"request",
                  "description":"request",
                  "required":true,
                  "schema":{
                     "$ref":"#/definitions/LoginRequest"
                  }
               }
            ],
            "responses":{
               "200":{
                  "description":"OK",
                  "schema":{
                     "$ref":"#/definitions/LoginResponse"
                  }
               },
               "201":{
                  "description":"Created"
               },
               "401":{
                  "description":"Unauthorized"
               },
               "403":{
                  "description":"Forbidden"
               },
               "404":{
                  "description":"Not Found"
               }
            }
         }
      }
   },
   "definitions":{
      "QuestionAnswerResponse":{
         "type":"object",
         "properties":{
            "answerMap":{
               "type":"object",
               "additionalProperties":{
                  "type":"string"
               }
            },
            "answerSelected":{
               "type":"integer",
               "format":"int32"
            },
            "correctAnswers":{
               "type":"integer",
               "format":"int32"
            },
            "maxQuestions":{
               "type":"integer",
               "format":"int32"
            },
            "question":{
               "type":"string"
            },
            "questionId":{
               "type":"integer",
               "format":"int32"
            },
            "requestId":{
               "type":"string"
            }
         }
      },
      "LoginResponse":{
         "type":"object",
         "properties":{
            "id":{
               "type":"string"
            },
            "requestId":{
               "type":"string"
            },
            "statusCode":{
               "type":"string"
            }
         }
      },
      "QuestionAnswerRequest":{
         "type":"object",
         "properties":{
            "answerId":{
               "type":"integer",
               "format":"int32"
            },
            "questionId":{
               "type":"integer",
               "format":"int32"
            },
            "requestId":{
               "type":"string"
            },
            "userId":{
               "type":"integer",
               "format":"int32"
            }
         }
      },
      "LoginRequest":{
         "type":"object",
         "properties":{
            "password":{
               "type":"string"
            },
            "requestId":{
               "type":"string"
            },
            "username":{
               "type":"string"
            }
         }
      }
   }
}